requirements.txt
fastapi
uvicorn
python-socketio
aiofiles
websockets
requests
customtkinter
pynacl
sqlalchemy
